# static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RosterIQ API",
    "description": "AI-powered predictive rostering for Australian hospitality venues. Manage venues, employees, rosters, forecasts, and compliance with intelligent optimization.",
    "version": "1.0.0",
    "contact": {
      "name": "RosterIQ Support",
      "url": "https://rosteriq.io",
      "email": "support@rosteriq.io"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://rosteriq.io/license"
    }
  },
  "servers": [
    {
      "url": "https://api.rosteriq.io",
      "description": "Production server"
    },
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from login endpoint"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for service-to-service authentication"
      }
    },
    "schemas": {
...
      "Shift": {
        "type": "object",
        "properties": {
          "shift_id": {
            "type": "string",
            "example": "shift_001"
          },
          "roster_id": {
            "type": "string",
            "example": "roster_abc123"
          },
          "employee_id": {
            "type": "string",
            "example": "emp_789xyz"
          },
          "date": {
            "type": "string",
            "format": "date",
            "example": "2026-05-01"
          },
          "start_time": {
            "type": "string",
            "format": "time",
            "example": "10:00"
          },
          "end_time": {
            "type": "string",
            "format": "time",
            "example": "18:00"
          },
          "duration_hours": {
            "type": "number",
            "example": 8
          },
          "status": {
            "type": "string",
            "enum": ["scheduled", "published", "worked", "cancelled"],
            "example": "scheduled"
          },
          "cost": {
            "type": "number",
            "example": 240.50
          }
        },
        "required": ["shift_id", "employee_id", "date", "start_time", "end_time"]
      },
      "Roster": {
        "type": "object",
        "properties": {
          "roster_id": {
            "type": "string",
            "example": "roster_abc123"
          },
          "venue_id": {
            "type": "string",
            "example": "venue_123abc"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "example": "2026-05-01"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "example": "2026-05-14"
          },
          "shifts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Shift"
            }
          },
          "shifts_count": {
            "type": "integer",
            "example": 42
          },
          "total_cost": {
            "type": "number",
            "example": 10080.50
          },
          "status": {
            "type": "string",
            "enum": ["draft", "published", "completed"],
            "example": "draft"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["roster_id", "venue_id", "start_date"]
      },
...
    }
  },
  "paths": {
...
    "/rosters/{roster_id}": {
      "get": {
        "tags": ["Rosters"],
        "summary": "Get roster details with all shifts",
        "operationId": "get_roster",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "roster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Roster with shifts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Roster"
                }
              }
            }
          },
          "404": {
            "description": "Roster not found"
          }
        }
      }
    },
...
    "/breaks/validate/{roster_id}": {
      "get": {
        "tags": ["Compliance"],
        "summary": "Validate break compliance for roster",
        "operationId": "validate_breaks",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "roster_id",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "Break validation results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "compliant": {"type": "boolean"},
                    "violations": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "shift_id": {"type": "string"},
                          "violation_type": {"type": "string"},
                          "message": {"type": "string"}
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}